changed lots of stuff as usually not breaking it down into nice commits.  For one, I changed taskpopulate.  Now it saves the file position in the datastore so that the position doesn't get evicted from memcache (resulting in repeated nodes).  In the process of doing that I created a handy little class called Property which allows you to save a String, Long, or Int in the datastore associated with a property name.  I think that will come in useful in the future (and due to objectify's memcache caching reading property values should be quick if they are read frequently).
on the client I've changed isLoaded from default on to default off.  And I've implemented a bunch of stuff to keep track of the isLoaded state of ViewNodes and to build lists of loaded ViewNodes.

And I probably did lots of other stuff that I don't remember at the moment...